Changed: rename File to FileModel
