+++ src/files/domain/repositories/IFilesRepository.ts
@@ -4,7 +4,7 @@ import { FileUserPermissions } from '../models/FileUserPermissions'
 import { FileSearchCriteria, FileOrderCriteria } from '../models/FileCriteria'
 import { FileCounts } from '../models/FileCounts'
 import { FileDownloadSizeMode } from '../models/FileDownloadSizeMode'
-import { File as FileModel } from '../models/File'
+import { FileModel } from '../models/FileModel'
 import { Dataset } from '../../../datasets'
 import { FileUploadDestination } from '../models/FileUploadDestination'
 import { UploadedFileDTO } from '../dtos/UploadedFileDTO'
