+++ src/files/domain/useCases/GetFile.ts
@@ -1,22 +1,22 @@
 import { IFilesRepository } from '../repositories/IFilesRepository'
-import { File } from '../models/File'
+import { FileModel } from '../models/FileModel'
 import { DatasetNotNumberedVersion } from '../../../datasets'
 import { UseCase } from '../../../core/domain/useCases/UseCase'
 
-export class GetFile implements UseCase<File> {
+export class GetFile implements UseCase<FileModel> {
   constructor(private readonly filesRepository: IFilesRepository) {}
 
   /**
-   * Returns a File instance, given the search parameters to identify it.
+   * Returns a FileModel instance, given the search parameters to identify it.
    *
    * @param {number | string} [fileId] - The File identifier, which can be a string (for persistent identifiers), or a number (for numeric identifiers).
    * @param {string | DatasetNotNumberedVersion} [datasetVersionId=DatasetNotNumberedVersion.LATEST] - The dataset version identifier, which can be a version-specific numeric string (for example, 1.0) or a DatasetNotNumberedVersion enum value. If this parameter is not set, the default value is: DatasetNotNumberedVersion.LATEST
-   * @returns {Promise<File>}
+   * @returns {Promise<FileModel>}
    */
   async execute(
     fileId: number | string,
     datasetVersionId: string | DatasetNotNumberedVersion = DatasetNotNumberedVersion.LATEST
-  ): Promise<File> {
-    return (await this.filesRepository.getFile(fileId, datasetVersionId, false)) as File
+  ): Promise<FileModel> {
+    return (await this.filesRepository.getFile(fileId, datasetVersionId, false)) as FileModel
   }
 }
